# spring-server/src/main/java/io/swagger/api/SubcategorieApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.27).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.SUBCATEGORIEITEM;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2022-05-21T14:27:53.479Z")

@Validated
@Api(value = "subcategorie", description = "the subcategorie API")
@RequestMapping(value = "/ords/{db}")
public interface SubcategorieApi {

    @ApiOperation(value = "", nickname = "subcategorieGet", notes = "Retrieve records from SUBCATEGORIE", response = Object.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "The queried record.", response = Object.class) })
    @RequestMapping(value = "/subcategorie",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Object> subcategorieGet(@PathVariable("db") String db);


    @ApiOperation(value = "", nickname = "subcategorieIdDelete", notes = "Remove a record from SUBCATEGORIE", response = SUBCATEGORIEITEM.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 204, message = "Deleted result.", response = SUBCATEGORIEITEM.class) })
    @RequestMapping(value = "/subcategorie/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.DELETE)
    ResponseEntity<SUBCATEGORIEITEM> subcategorieIdDelete(@PathVariable("db") String db, @Pattern(regexp="^[^/]+$") @ApiParam(value = "implicit",required=true) @PathVariable("id") String id);


    @ApiOperation(value = "", nickname = "subcategorieIdGet", notes = "Retrieve a record from SUBCATEGORIE", response = SUBCATEGORIEITEM.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "The queried record.", response = SUBCATEGORIEITEM.class) })
    @RequestMapping(value = "/subcategorie/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<SUBCATEGORIEITEM> subcategorieIdGet(@PathVariable("db") String db, @Pattern(regexp="^[^/]+$") @ApiParam(value = "implicit",required=true) @PathVariable("id") String id);


    @ApiOperation(value = "", nickname = "subcategorieIdPut", notes = "Create or update a record on SUBCATEGORIE", response = SUBCATEGORIEITEM.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "The successfully updated record.", response = SUBCATEGORIEITEM.class),
        @ApiResponse(code = 201, message = "The successfully created record.", response = SUBCATEGORIEITEM.class) })
    @RequestMapping(value = "/subcategorie/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.PUT)
    ResponseEntity<SUBCATEGORIEITEM> subcategorieIdPut(@PathVariable("db") String db, @Pattern(regexp="^[^/]+$") @ApiParam(value = "implicit",required=true) @PathVariable("id") String id,@ApiParam(value = "" ,required=true )  @Valid @RequestBody SUBCATEGORIEITEM payload);


    @ApiOperation(value = "", nickname = "subcategoriePost", notes = "Create a new record on SUBCATEGORIE", response = SUBCATEGORIEITEM.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "The successfully created record.", response = SUBCATEGORIEITEM.class) })
    @RequestMapping(value = "/subcategorie",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<SUBCATEGORIEITEM> subcategoriePost(@PathVariable("db") String db, @ApiParam(value = "" ,required=true )  @Valid @RequestBody SUBCATEGORIEITEM payload);

}